Add --open-browser flag to daml start
I went with the yes/no/auto approach we already use in the assistant
and moved the logic for that to da-hs-base.
