+++ daml-assistant/src/DAML/Assistant/Command.hs
@@ -18,7 +18,7 @@ import DAML.Assistant.Types
 import Data.List
 import Data.Maybe
 import Data.Foldable
-import Options.Applicative
+import Options.Applicative.Extended
 
 getCommand :: [SdkCommandInfo] -> IO Command
 getCommand sdkCommands =
@@ -64,18 +64,6 @@ installParser = InstallOptions
     <*> iflag ActivateInstall "activate" mempty "Activate installed version of daml"
     <*> iflag ForceInstall "force" (short 'f') "Overwrite existing installation"
     <*> iflag QuietInstall "quiet" (short 'q') "Don't display installation messages"
-    <*> fmap SetPath (flagWithAuto "set-path" True "Adjust PATH automatically. This option only has an effect on Windows.")
+    <*> fmap SetPath (flagYesNoAuto "set-path" True "Adjust PATH automatically. This option only has an effect on Windows.")
     where
         iflag p name opts desc = fmap p (switch (long name <> help desc <> opts))
-
--- | This constructs flags that can be set to yes, no, or auto to control a boolean value
--- with auto using the default.
-flagWithAuto :: String -> Bool -> String -> Parser Bool
-flagWithAuto flagName defaultValue helpText =
-    option reader (long flagName <> value defaultValue <> help (helpText <> commonHelp))
-  where reader = eitherReader $ \case
-            "yes" -> Right True
-            "no" -> Right False
-            "auto" -> Right defaultValue
-            s -> Left ("Expected \"yes\", \"no\" or \"auto\" but got " <> show s)
-        commonHelp = " Can be set to \"yes\", \"no\" or \"auto\" to select the default (" <> show defaultValue <> ")"
